third_party/WebKit/LayoutTests/vr/latest/navigator_vr_blocked_by_getVRDisplays.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

promise_test(t => navigator.getVRDisplays().then(() => {
  assert_true(navigator.vr == null);
}), "Test that access to navigator.vr is blocked once navigator.getVRDisplays has been called.");

</script>
